feat: main features done

+++ common/src/main/java/ru/practicum/UniqueElementsValidator.java
@@ -6,13 +6,13 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-public class UniqueElementsValidator implements ConstraintValidator<UniqueElements, List<Integer>> {
+public class UniqueElementsValidator implements ConstraintValidator<UniqueElements, List<Long>> {
     @Override
-    public boolean isValid(List<Integer> value, ConstraintValidatorContext context) {
+    public boolean isValid(List<Long> value, ConstraintValidatorContext context) {
         if (value == null) {
             return true; // Null lists are considered valid. Use @NotNull for non-null validation.
         }
-        Set<Integer> uniqueElements = new HashSet<>(value);
+        Set<Long> uniqueElements = new HashSet<>(value);
         return uniqueElements.size() == value.size();
     }
 }
